+++ assets/js/blocks/reviews/reviews-by-product/edit.js
@@ -10,24 +10,19 @@ import {
 	Button,
 	PanelBody,
 	Placeholder,
-	Spinner,
 	Toolbar,
 	withSpokenMessages,
 } from '@wordpress/components';
 import { SearchListItem } from '@woocommerce/components';
 import { Fragment } from '@wordpress/element';
-import { compose } from '@wordpress/compose';
-import { escapeHTML } from '@wordpress/escape-html';
 import PropTypes from 'prop-types';
 
 /**
  * Internal dependencies
  */
-import ApiErrorPlaceholder from '../../../components/api-error-placeholder';
 import EditorBlock from './editor-block.js';
 import ProductControl from '../../../components/product-control';
 import { IconReviewsByProduct } from '../../../components/icons';
-import { withProduct } from '../../../hocs';
 import { ENABLE_REVIEW_RATING, SHOW_AVATARS } from '../../../constants';
 import { getSharedReviewContentControls, getSharedReviewListControls } from '../edit.js';
 import { getBlockClassName } from '../utils.js';
@@ -35,7 +30,7 @@ import { getBlockClassName } from '../utils.js';
 /**
  * Component to handle edit mode of "Reviews by Product".
  */
-const ReviewsByProductEditor = ( { attributes, debouncedSpeak, error, getProduct, isLoading, product, setAttributes } ) => {
+const ReviewsByProductEditor = ( { attributes, debouncedSpeak, setAttributes } ) => {
 	attributes.showReviewImage = ( SHOW_AVATARS || attributes.imageType === 'product' ) && attributes.showReviewImage;
 	attributes.showReviewRating = ENABLE_REVIEW_RATING && attributes.showReviewRating;
 
@@ -112,27 +107,6 @@ const ReviewsByProductEditor = ( { attributes, debouncedSpeak, error, getProduct
 		);
 	};
 
-	const renderApiError = () => (
-		<ApiErrorPlaceholder
-			className="wc-block-featured-product-error"
-			error={ error }
-			isLoading={ isLoading }
-			onRetry={ getProduct }
-		/>
-	);
-
-	const renderLoadingScreen = () => {
-		return (
-			<Placeholder
-				icon={ <IconReviewsByProduct className="block-editor-block-icon" /> }
-				label={ __( 'Reviews by Product', 'woo-gutenberg-products-block' ) }
-				className="wc-block-reviews-by-product"
-			>
-				<Spinner />
-			</Placeholder>
-		);
-	};
-
 	const renderEditMode = () => {
 		const onDone = () => {
 			setAttributes( { editMode: false } );
@@ -187,31 +161,9 @@ const ReviewsByProductEditor = ( { attributes, debouncedSpeak, error, getProduct
 		);
 	};
 
-	const renderNoReviews = () => (
-		<Placeholder
-			className="wc-block-reviews-by-product"
-			icon={ <IconReviewsByProduct className="block-editor-block-icon" /> }
-			label={ __( 'Reviews by Product', 'woo-gutenberg-products-block' ) }
-		>
-			<div dangerouslySetInnerHTML={ {
-				__html: sprintf(
-					__(
-						"This block lists reviews for a selected product. %s doesn't have any reviews yet, but they will show up here when it does.",
-						'woo-gutenberg-products-block'
-					),
-					'<strong>' + escapeHTML( product.name ) + '</strong>'
-				),
-			} } />
-		</Placeholder>
-	);
-
 	const renderViewMode = () => {
 		if ( ! showReviewContent && ! showReviewRating && ! showReviewDate && ! showReviewerName && ! showReviewImage ) {
 			return renderHiddenContentPlaceholder();
-		}
-
-		if ( product.review_count === 0 ) {
-			return renderNoReviews();
 		}
 
 		return (
@@ -221,16 +173,8 @@ const ReviewsByProductEditor = ( { attributes, debouncedSpeak, error, getProduct
 		);
 	};
 
-	if ( error ) {
-		return renderApiError();
-	}
-
 	if ( ! productId || editMode ) {
 		return renderEditMode();
-	}
-
-	if ( ! product || isLoading ) {
-		return renderLoadingScreen();
 	}
 
 	return (
@@ -255,19 +199,8 @@ ReviewsByProductEditor.propTypes = {
 	 * A callback to update attributes.
 	 */
 	setAttributes: PropTypes.func.isRequired,
-	// from withProduct
-	error: PropTypes.object,
-	getProduct: PropTypes.func,
-	isLoading: PropTypes.bool,
-	product: PropTypes.shape( {
-		name: PropTypes.node,
-		review_count: PropTypes.number,
-	} ),
 	// from withSpokenMessages
 	debouncedSpeak: PropTypes.func.isRequired,
 };
 
-export default compose( [
-	withProduct,
-	withSpokenMessages,
-] )( ReviewsByProductEditor );
+export default withSpokenMessages( ReviewsByProductEditor );
